ESP32 analog IO Example
Example code that given a potentiometer input to pin 36 of the esp32 pico, outputs a pwm signal at pin 4 who's dutycycle is equal to the read analog value. If pin 4 is connected to an led and series resistor (Lab 0) then the resistor should dim and brighten with the potentiometer angle.