ESP32 Digital IO Example
Example code that reads a digital input at pin 7 and writes the read value to pin 4. If a button is setup on pin 7 and pin 7 is set to be input pullup, and if an led and series resistor are setup on pin 4 then when the button is depressed the led will turn off and when the button is released the led will turn on.